Done refactoring data package. Next, update the package info
