+++ someAutoCADColumns.txt
@@ -1,0 +1,42 @@
+# These are just a bunch of columns AutoCAD Excel files may contain.
+# I'll hold onto this for now, but I likely won't need this once I've abstractified grading criteria
+Count
+Name
+Angle
+Center X
+Center Y
+Center Z
+Color
+Delta X
+Delta Y
+Delta Z
+Diameter
+Drawing
+End X
+End Y
+End Z
+Layer
+Length
+Linetype
+Linetype Scale
+Lineweight
+Plot Style
+Start X
+Start Y
+Start Z
+Thickness
+Contents
+ContentsRTF
+Position X
+Position Y
+Position Z
+Rotation
+ShowBorders
+Width
+Area
+Closed
+Global Width
+Dim Style
+DynamicDimension
+TextDefinedSize
+Value
