+++ src/main/java/autocadDrawingChecker/grading/criteria/AbstractElementCriteria.java
@@ -1,7 +1,7 @@
 package autocadDrawingChecker.grading.criteria;
 
-import autocadDrawingChecker.data.elements.AutoCADElement;
-import autocadDrawingChecker.data.elements.AutoCADExport;
+import autocadDrawingChecker.data.AutoCADElement;
+import autocadDrawingChecker.data.AutoCADExport;
 import autocadDrawingChecker.grading.AutoCADElementMatcher;
 import autocadDrawingChecker.grading.MatchingAutoCADElements;
 import java.util.Arrays;
