+++ src/main/java/autocadDrawingChecker/grading/criteria/AbstractGradingCriteria.java
@@ -1,6 +1,6 @@
 package autocadDrawingChecker.grading.criteria;
 
-import autocadDrawingChecker.data.elements.AutoCADExport;
+import autocadDrawingChecker.data.AutoCADExport;
 
 /**
  * The AbstractGradingCriteria interface is used to
