+++ src/main/java/autocadDrawingChecker/grading/criteria/implementations/LineEnd.java
@@ -1,6 +1,6 @@
 package autocadDrawingChecker.grading.criteria.implementations;
 
-import autocadDrawingChecker.data.elements.AutoCADElement;
+import autocadDrawingChecker.data.AutoCADElement;
 
 /**
  *
