+++ src/main/java/autocadDrawingChecker/grading/criteria/implementations/TextMatches.java
@@ -1,6 +1,6 @@
 package autocadDrawingChecker.grading.criteria.implementations;
 
-import autocadDrawingChecker.data.elements.AutoCADElement;
+import autocadDrawingChecker.data.AutoCADElement;
 import autocadDrawingChecker.grading.criteria.AbstractElementCriteria;
 
 /**
